+++ backend/internal/routers.go
@@ -20,6 +20,6 @@ func SetRoutes(router *gin.Engine) {
 	v1 := router.Group("/v1")
 	{
 		v1.GET("/users/:user_id", controllers.GetUsersByIdControlller)
-		v1.PUT("/status", controllers.PutStatusHandlerFunc)
+		v1.PUT("/status", controllers.PutStatusController)
 	}
 }
