--- backend/internal/services/status.go
@@ -1,46 +1,0 @@
-package services
-
-import (
-	"fmt"
-
-	"github.com/ISDL-dev/ISDL-Sentinel/backend/internal/model"
-	"github.com/ISDL-dev/ISDL-Sentinel/backend/internal/repositories"
-	"github.com/ISDL-dev/ISDL-Sentinel/backend/internal/schema"
-)
-
-func SelectApplyStatusId(status string) (id int32, statusName string) {
-	var getStatusId int32 
-	var getStatusName string
-	var targetStatusName string
-	getRows, err := repositories.GetAllStatusRepository()
-	if err != nil {
-		return 0, targetStatusName
-	}
-	if (model.IsInRoom(status)){
-		targetStatusName = model.OUT_ROOM
-	} else {
-		targetStatusName = model.IN_ROOM
-	}
-	for getRows.Next() {
-		err := getRows.Scan(&getStatusId, &getStatusName)
-		if err != nil {
-			return 0, targetStatusName
-		}
-		if getStatusName == targetStatusName{
-			return getStatusId, targetStatusName 
-		}
-	}
-	return 0, targetStatusName
-}
-
-func PutStatusService(status schema.Status) (user schema.Status, err error) {
-	var applyStatusId int32
-	var targetStatusName string
-	applyStatusId, targetStatusName = SelectApplyStatusId(status.Status)
-	err = repositories.PutStatusRepository(status.UserId, applyStatusId)
-	if err != nil {
-		return schema.Status{}, fmt.Errorf("failed to change user status: %v", err)
-	}
-
-	return schema.Status{UserId: status.UserId, Status: targetStatusName}, nil
-}
